+++ src/pages/Home/routes/Home.jsx
@@ -1,50 +1,49 @@
 /* eslint-disable no-nested-ternary */
-import React, { useEffect, useState } from 'react';
-import { throttle } from 'lodash';
+import React, { useState } from 'react';
 import Header from '../../../components/Element/Header/Header';
 import Footer from '../../../components/Element/Navbar/Navbar';
 import '../style/Home.scss';
 import { HomeWithPost } from '../components/HomeWithPost';
 import { HomeWithoutPost } from '../components/HomeWithoutPost';
-import { fetchClosure } from '../api';
-import PostSkeleton from '../../../components/Element/Post/PostSkeleton';
-import { getScrollTop, getDocumentHeight } from '../../../utils/scroll';
+import PostSkeleton from '../../../components/Skeleton/PostSkeleton';
+import { scrollHook } from '../../../hooks/scroll';
 
 export function Home() {
   const [posts, setPosts] = useState([]);
   const [isLoading, setIsLoading] = useState(true);
   const [hasMoreData, setHasMoreData] = useState(true);
-  const getFeedData = fetchClosure(4);
+  scrollHook({ url: '/post/feed', cnt: 4, setData: setPosts, setIsLoading, hasMoreData, setHasMoreData });
+  // const getFeedData = fetchClosure('/post/feed', 4);
 
-  useEffect(() => {
-    (async () => {
-      const res = await getFeedData();
-      setPosts(res.data.posts);
+  // useEffect(() => {
+  //   (async () => {
+  //     const res = await getFeedData();
+  //     setPosts(res.data.posts);
 
-      setIsLoading(false);
-    })();
-  }, []);
+  //     setIsLoading(false);
+  //   })();
+  // }, []);
 
-  const handleScroll = async () => {
-    if (!hasMoreData) return;
+  // const handleScroll = async () => {
+  //   if (!hasMoreData) return;
 
-    if (Math.ceil(getScrollTop()) >= getDocumentHeight() - window.innerHeight) {
-      const res = await getFeedData();
-      if (res.data.posts.length === 0) {
-        setHasMoreData(false);
-      } else {
-        setPosts(prev => prev.concat(res.data.posts));
-      }
-    }
-  };
+  //   if (Math.ceil(getScrollTop()) >= getDocumentHeight() - window.innerHeight) {
+  //     const res = await getFeedData();
+  //     if (res.data.posts.length === 0) {
+  //       setHasMoreData(false);
+  //     } else {
+  //       setPosts(prev => prev.concat(res.data.posts));
+  //     }
+  //   }
+  // };
 
-  useEffect(() => {
-    const throttledScroll = throttle(handleScroll, 500);
-    window.addEventListener('scroll', throttledScroll);
-    return () => {
-      window.removeEventListener('scroll', throttledScroll);
-    };
-  }, []);
+  // useEffect(() => {
+  //   const throttledScroll = throttle(handleScroll, 500);
+  //   window.addEventListener('scroll', throttledScroll);
+  //   return () => {
+  //     window.removeEventListener('scroll', throttledScroll);
+  //   };
+  // }, []);
 
   return (
     <>
